+++ pkg/platform/resource_permission_test.go
@@ -1,3 +1,17 @@
+// Copyright (c) JFrog Ltd. (2025)
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
 package platform_test
 
 import (
@@ -1263,6 +1277,95 @@ func TestAccPermission_build_multiple_targets_validation(t *testing.T) {
 			{
 				Config:      config,
 				ExpectError: regexp.MustCompile(`.*Attribute.*targets.*must contain at most.*1.*elements.*`),
+			},
+		},
+	})
+}
+
+func TestAccPermission_build_custom_repository(t *testing.T) {
+	_, fqrn, permissionName := testutil.MkNames("test-permission", "platform_permission")
+	_, _, userName := testutil.MkNames("test-user", "artifactory_managed_user")
+
+	temp := `
+	resource "artifactory_managed_user" "{{ .userName }}" {
+		name = "{{ .userName }}"
+		email = "{{ .userName }}@tempurl.org"
+		password = "Password!123"
+	}
+
+	resource "platform_permission" "{{ .name }}" {
+		name = "{{ .name }}"
+
+		build = {
+			actions = {
+				users = [
+					{
+						name = artifactory_managed_user.{{ .userName }}.name
+						permissions = ["READ"]
+					}
+				]
+			}
+
+			targets = [
+				{
+					name = "{{ .buildRepoName }}"
+					include_patterns = ["**"]
+				}
+			]
+		}
+	}`
+
+	testData := map[string]string{
+		"name":          permissionName,
+		"userName":      userName,
+		"buildRepoName": "custom-build-info",
+	}
+
+	config := util.ExecuteTemplate(permissionName, temp, testData)
+
+	updatedTestData := map[string]string{
+		"name":          permissionName,
+		"userName":      userName,
+		"buildRepoName": "another-custom-build-info",
+	}
+	updatedConfig := util.ExecuteTemplate(permissionName, temp, updatedTestData)
+
+	resource.Test(t, resource.TestCase{
+		PreCheck:                 func() { testAccPreCheck(t) },
+		ProtoV6ProviderFactories: testAccProviders(),
+		ExternalProviders: map[string]resource.ExternalProvider{
+			"artifactory": {
+				Source: "jfrog/artifactory",
+			},
+		},
+		// CheckDestroy: testAccCheckPermissionDestroy(fqrn),
+		Steps: []resource.TestStep{
+			{
+				Config: config,
+				Check: resource.ComposeTestCheckFunc(
+					resource.TestCheckResourceAttr(fqrn, "name", testData["name"]),
+					resource.TestCheckResourceAttr(fqrn, "build.targets.#", "1"),
+					resource.TestCheckResourceAttr(fqrn, "build.targets.0.name", testData["buildRepoName"]),
+					resource.TestCheckResourceAttr(fqrn, "build.targets.0.include_patterns.#", "1"),
+					resource.TestCheckTypeSetElemAttr(fqrn, "build.targets.0.include_patterns.*", "**"),
+				),
+			},
+			{
+				Config: updatedConfig,
+				Check: resource.ComposeTestCheckFunc(
+					resource.TestCheckResourceAttr(fqrn, "name", updatedTestData["name"]),
+					resource.TestCheckResourceAttr(fqrn, "build.targets.#", "1"),
+					resource.TestCheckResourceAttr(fqrn, "build.targets.0.name", updatedTestData["buildRepoName"]),
+					resource.TestCheckResourceAttr(fqrn, "build.targets.0.include_patterns.#", "1"),
+					resource.TestCheckTypeSetElemAttr(fqrn, "build.targets.0.include_patterns.*", "**"),
+				),
+			},
+			{
+				ResourceName:                         fqrn,
+				ImportState:                          true,
+				ImportStateId:                        permissionName,
+				ImportStateVerify:                    true,
+				ImportStateVerifyIdentifierAttribute: "name",
 			},
 		},
 	})